Add files via upload
Structure of european options on an underlying asset to guarantee the payoff to our benefit, but if  the extreme values are achieved lost will be enormous. I studied the underlying asset and it never reached so extreme values, so profiit will be made unless a  huge catastrophe occurs, which seems impossible.